+++ src/main/java/org/seedstack/feign/internal/FeignErrorCode.java
@@ -1,5 +1,5 @@
-/**
- * Copyright (c) 2013-2016, The SeedStack authors <http://seedstack.org>
+/*
+ * Copyright © 2013-2019, The SeedStack authors <http://seedstack.org>
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -9,14 +9,14 @@ package org.seedstack.feign.internal;
 
 import org.seedstack.shed.exception.ErrorCode;
 
-enum FeignErrorCode implements ErrorCode {
-    ERROR_BUILDING_HYSTRIX_CLIENT,
+public enum FeignErrorCode implements ErrorCode {
+    BAD_TARGET_CLASS,
+    BAD_FALLBACK_CLASS,
     ERROR_INSTANTIATING_CONTRACT,
     ERROR_INSTANTIATING_DECODER,
     ERROR_INSTANTIATING_ENCODER,
     ERROR_INSTANTIATING_TARGET,
-    ERROR_INSTANTIATING_TARGET_BAD_TARGET_CLASS,
     ERROR_INSTANTIATING_FALLBACK,
     ERROR_INSTANTIATING_LOGGER,
-    HYSTRIX_NOT_PRESENT 
+    HYSTRIX_NOT_PRESENT
 }