+++ src/main/java/org/seedstack/feign/internal/FeignModule.java
@@ -1,5 +1,5 @@
-/**
- * Copyright (c) 2013-2016, The SeedStack authors <http://seedstack.org>
+/*
+ * Copyright © 2013-2019, The SeedStack authors <http://seedstack.org>
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -7,42 +7,30 @@
  */
 package org.seedstack.feign.internal;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-
 import com.google.inject.AbstractModule;
-import com.google.inject.multibindings.Multibinder;
-
-import feign.Target;
+import java.util.Collection;
+import javax.net.ssl.SSLContext;
 
 class FeignModule extends AbstractModule {
-    private final Collection<Class<?>> feignApis;
-    private final List<Class<? extends Target<?>>> feignTargets = new ArrayList<>();
+    private final Collection<Class<?>> feignInterfaces;
+    private final Collection<Class<?>> bindings;
+    private final SSLContext sslContext;
 
-    FeignModule(Collection<Class<?>> feignApis, Collection<Class<?>> targetClasses) {
-        this.feignApis = feignApis;
-        resolveFeignTargets(targetClasses);
-
-    }
-
-    @SuppressWarnings("unchecked")
-    private void resolveFeignTargets(Collection<Class<?>> targetClasses) {
-        targetClasses.stream()
-                .map(x -> (Class<? extends Target<?>>) x)
-                .forEach(feignTargets::add);
+    FeignModule(Collection<Class<?>> feignInterfaces, Collection<Class<?>> bindings, SSLContext sslContext) {
+        this.feignInterfaces = feignInterfaces;
+        this.bindings = bindings;
+        this.sslContext = sslContext;
     }
 
     @Override
-    @SuppressWarnings({ "unchecked", "rawtypes" })
+    @SuppressWarnings("unchecked")
     protected void configure() {
-        Multibinder<Target> targetMultibinder = Multibinder.newSetBinder(binder(), Target.class);
-        for (Class<? extends Target> targetClass : feignTargets) {
-            targetMultibinder.addBinding().to((Class<? extends Target>) targetClass);
+        for (Class<?> binding : bindings) {
+            bind(binding);
         }
 
-        for (Class<?> feignApi : feignApis) {
-            bind(feignApi).toProvider((javax.inject.Provider) new FeignProvider(feignApi));
+        for (Class<?> feignApi : feignInterfaces) {
+            bind(feignApi).toProvider(new FeignProvider(feignApi, sslContext));
         }
     }