+++ src/test/java/org/seedstack/feign/fixtures/TestContract.java
@@ -1,5 +1,5 @@
-/**
- * Copyright (c) 2013-2016, The SeedStack authors <http://seedstack.org>
+/*
+ * Copyright © 2013-2019, The SeedStack authors <http://seedstack.org>
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -10,13 +10,6 @@ package org.seedstack.feign.fixtures;
 import static feign.Util.checkState;
 import static feign.Util.emptyToNull;
 
-import java.lang.annotation.Annotation;
-import java.lang.reflect.Method;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.LinkedHashMap;
-import java.util.Map;
-
 import feign.Body;
 import feign.Contract;
 import feign.HeaderMap;
@@ -24,7 +17,15 @@ import feign.Headers;
 import feign.MethodMetadata;
 import feign.Param;
 import feign.QueryMap;
+import feign.Request;
 import feign.RequestLine;
+import java.lang.annotation.Annotation;
+import java.lang.reflect.Method;
+import java.nio.charset.Charset;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.LinkedHashMap;
+import java.util.Map;
 
 public class TestContract extends Contract.BaseContract {
 
@@ -33,8 +34,41 @@ public class TestContract extends Contract.BaseContract {
     public static boolean hasBeenUsed() {
         return used;
     }
+
     private static void use() {
         used = true;
+    }
+
+    private static <K, V> boolean searchMapValuesContainsSubstring(Map<K, Collection<String>> map,
+            String search) {
+        Collection<Collection<String>> values = map.values();
+        if (values == null) {
+            return false;
+        }
+
+        for (Collection<String> entry : values) {
+            for (String value : entry) {
+                if (value.indexOf(search) != -1) {
+                    return true;
+                }
+            }
+        }
+
+        return false;
+    }
+
+    private static Map<String, Collection<String>> toMap(String[] input) {
+        Map<String, Collection<String>> result = new LinkedHashMap<String, Collection<String>>(
+                input.length);
+        for (String header : input) {
+            int colon = header.indexOf(':');
+            String name = header.substring(0, colon);
+            if (!result.containsKey(name)) {
+                result.put(name, new ArrayList<String>(1));
+            }
+            result.get(name).add(header.substring(colon + 2));
+        }
+        return result;
     }
 
     // Copycat of DefaultContract with test-flag
@@ -65,16 +99,16 @@ public class TestContract extends Contract.BaseContract {
                 checkState(requestLine.indexOf('/') == -1,
                         "RequestLine annotation didn't start with an HTTP verb on method %s.",
                         method.getName());
-                data.template().method(requestLine);
+                data.template().method(Request.HttpMethod.valueOf(requestLine));
                 return;
             }
-            data.template().method(requestLine.substring(0, requestLine.indexOf(' ')));
+            data.template().method(Request.HttpMethod.valueOf(requestLine.substring(0, requestLine.indexOf(' '))));
             if (requestLine.indexOf(' ') == requestLine.lastIndexOf(' ')) {
                 // no HTTP version is ok
-                data.template().append(requestLine.substring(requestLine.indexOf(' ') + 1));
+                data.template().uri(requestLine.substring(requestLine.indexOf(' ') + 1));
             } else {
                 // skip HTTP version
-                data.template().append(
+                data.template().uri(
                         requestLine.substring(requestLine.indexOf(' ') + 1,
                                 requestLine.lastIndexOf(' ')));
             }
@@ -85,11 +119,7 @@ public class TestContract extends Contract.BaseContract {
             String body = Body.class.cast(methodAnnotation).value();
             checkState(emptyToNull(body) != null, "Body annotation was empty on method %s.",
                     method.getName());
-            if (body.indexOf('{') == -1) {
-                data.template().body(body);
-            } else {
-                data.template().bodyTemplate(body);
-            }
+            data.template().body(Request.Body.bodyTemplate(body, Charset.defaultCharset()));
         } else if (annotationType == Headers.class) {
             String[] headersOnMethod = Headers.class.cast(methodAnnotation).value();
             checkState(headersOnMethod.length > 0, "Headers annotation was empty on method %s.",
@@ -138,38 +168,6 @@ public class TestContract extends Contract.BaseContract {
             }
         }
         return isHttpAnnotation;
-    }
-
-    private static <K, V> boolean searchMapValuesContainsSubstring(Map<K, Collection<String>> map,
-            String search) {
-        Collection<Collection<String>> values = map.values();
-        if (values == null) {
-            return false;
-        }
-
-        for (Collection<String> entry : values) {
-            for (String value : entry) {
-                if (value.indexOf(search) != -1) {
-                    return true;
-                }
-            }
-        }
-
-        return false;
-    }
-
-    private static Map<String, Collection<String>> toMap(String[] input) {
-        Map<String, Collection<String>> result = new LinkedHashMap<String, Collection<String>>(
-                input.length);
-        for (String header : input) {
-            int colon = header.indexOf(':');
-            String name = header.substring(0, colon);
-            if (!result.containsKey(name)) {
-                result.put(name, new ArrayList<String>(1));
-            }
-            result.get(name).add(header.substring(colon + 2));
-        }
-        return result;
     }
 
 }