+++ src/test/java/org/seedstack/feign/fixtures/TestFallback.java
@@ -1,5 +1,5 @@
-/**
- * Copyright (c) 2013-2016, The SeedStack authors <http://seedstack.org>
+/*
+ * Copyright © 2013-2019, The SeedStack authors <http://seedstack.org>
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -7,12 +7,18 @@
  */
 package org.seedstack.feign.fixtures;
 
+import org.seedstack.feign.fixtures.apis.HystrixEnabledAPI;
 import org.seedstack.feign.fixtures.apis.TestAPI;
 
-public class TestFallback implements TestAPI {
+public class TestFallback implements TestAPI, HystrixEnabledAPI {
     @Override
     public Message getMessage() {
         return new Message("Fallback response", "fallback");
+    }
+
+    @Override
+    public Message getProtectedMessage(String credentials) {
+        return new Message("Fallback protected response", "fallback");
     }
 
     @Override