+++ src/test/java/org/seedstack/feign/fixtures/apis/TargetableAPI.java
@@ -1,5 +1,5 @@
-/**
- * Copyright (c) 2013-2016, The SeedStack authors <http://seedstack.org>
+/*
+ * Copyright © 2013-2019, The SeedStack authors <http://seedstack.org>
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -7,14 +7,12 @@
  */
 package org.seedstack.feign.fixtures.apis;
 
+import feign.RequestLine;
 import org.seedstack.feign.FeignApi;
 import org.seedstack.feign.fixtures.Message;
 
-import feign.RequestLine;
-
-
 @FeignApi
-public interface TestContractAPI {
+public interface TargetableAPI {
 
     @RequestLine("GET /message")
     Message getMessage();